HSEARCH-3256 Move DslConverter enum to spi package
